lib/pages/requirements.txt
google-generativeai
fastapi
pydantic
python-dotenv
uvicorn

